requirements.txt
discord.py
psycopg2_binary==2.9.6
python-dotenv==1.0.0
